# docs/swagger.yaml
basePath: /
definitions:
...
  dtos.CardInput:
    properties:
      closing_offset_days:
        description: 'Opcional, padrão: 7'
        example: 7
        maximum: 31
        minimum: 1
        type: integer
      due_day:
        example: 10
        maximum: 31
        minimum: 1
        type: integer
      name:
        example: Nubank Platinum
        type: string
    type: object
  dtos.CardOutput:
    properties:
      closing_offset_days:
        description: 'Quantos dias antes do vencimento fecha (padrão brasileiro: 7)'
        example: 7
        type: integer
      created_at:
        example: "2025-01-15T10:30:00Z"
        type: string
      due_day:
        example: 10
        type: integer
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Nubank Platinum
        type: string
      updated_at:
        example: "2025-01-20T08:00:00Z"
        type: string
    type: object
  dtos.CardPaginatedOutput:
    properties:
      data:
        items:
          $ref: '#/definitions/dtos.CardOutput'
        type: array
      pagination:
        $ref: '#/definitions/dtos.CardPaginationMeta'
    type: object
  dtos.CardPaginationMeta:
    properties:
      has_next:
        example: true
        type: boolean
      limit:
        example: 20
        type: integer
      next_cursor:
        example: eyJmIjp7Im5hbWUiOiJOdWJhbmsifX0
        type: string
    type: object
...
  httperrors.ProblemDetail:
    properties:
      detail:
        type: string
      errors:
        additionalProperties: {}
        type: object
      instance:
        type: string
      request_id:
        type: string
      status:
        type: integer
      timestamp:
        type: string
      title:
        type: string
      trace_id:
        type: string
      type:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Jailton Junior
    url: https://github.com/jailtonjunior94
  description: |-
    API para gestão financeira pessoal com suporte a cartões de crédito, transações,
    orçamentos mensais, categorias e faturas com parcelamento.

    ## Valores Monetários
    Todos os valores monetários são representados como **strings decimais** (ex: `"1234.56"`)
    para preservar precisão e evitar erros de ponto flutuante.

    ## Paginação
    Endpoints de listagem usam **cursor-based pagination**. Parâmetros:
    - `limit` (int, default: 20, max: 100)
    - `cursor` (string, opcional): valor de `pagination.next_cursor` da resposta anterior

    ## Autenticação
    Endpoints protegidos requerem o header `Authorization: Bearer {token}`.
    Obtenha o token em `POST /api/v1/token`.
  license:
    name: MIT
  title: Financial API
  version: 1.0.0
paths:
...
  /api/v1/cards:
    get:
      description: Retorna a lista paginada de cartões do usuário autenticado (cursor-based
        pagination).
      parameters:
      - default: 20
        description: 'Itens por página (default: 20, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Cursor de paginação (retornado em pagination.next_cursor)
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista paginada de cartões
          schema:
            $ref: '#/definitions/dtos.CardPaginatedOutput'
        "400":
          description: Parâmetro inválido
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Listar cartões
      tags:
      - cards
    post:
      consumes:
      - application/json
      description: |-
        Cria um novo cartão de crédito para o usuário autenticado.
        O campo `closing_offset_days` é opcional (default: 7 dias antes do vencimento).
        - `due_day`: dia do mês em que a fatura vence (1–31)
        - `closing_offset_days`: quantos dias antes do vencimento a fatura fecha (1–31)
      parameters:
      - description: Dados do cartão
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CardInput'
      produces:
      - application/json
      responses:
        "201":
          description: Cartão criado com sucesso
          schema:
            $ref: '#/definitions/dtos.CardOutput'
        "400":
          description: Dados inválidos
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Criar cartão
      tags:
      - cards
  /api/v1/cards/{id}:
    delete:
      description: Remove um cartão do usuário autenticado. Esta operação é irreversível.
      parameters:
      - description: ID do cartão
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Cartão removido com sucesso
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "404":
          description: Cartão não encontrado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Remover cartão
      tags:
      - cards
    get:
      description: Retorna os detalhes de um cartão específico do usuário autenticado.
      parameters:
      - description: ID do cartão
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Dados do cartão
          schema:
            $ref: '#/definitions/dtos.CardOutput'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "404":
          description: Cartão não encontrado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Buscar cartão por ID
      tags:
      - cards
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um cartão existente do usuário autenticado.
      parameters:
      - description: ID do cartão
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados do cartão
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CardInput'
      produces:
      - application/json
      responses:
        "200":
          description: Cartão atualizado com sucesso
          schema:
            $ref: '#/definitions/dtos.CardOutput'
        "400":
          description: Dados inválidos
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "404":
          description: Cartão não encontrado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Atualizar cartão
      tags:
      - cards
...
securityDefinitions:
  BearerAuth:
    description: 'Token JWT no formato: `Bearer {token}`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"